test(Purchase): 구입 금액에 대한 예외 처리 테스트 이동
LottiesTest에서 PurchaseTest로 이동
