+++ src/test/java/lotto/domain/PurchaseTest.java
@@ -1,0 +1,66 @@
+package lotto.domain;
+
+import lotto.exception.InputException;
+import lotto.view.InputView;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+
+import java.io.ByteArrayInputStream;
+import java.io.InputStream;
+
+public class PurchaseTest {
+
+    private Purchase purchase;
+
+    @BeforeEach
+    void init() {
+        int purchaseCount = 2000;
+        purchase = Purchase.create(purchaseCount);
+    }
+
+    @DisplayName("구입 금액을 로또 하나 금액으로 나눠 총 몇개를 발행할지 구한다.")
+    @Test
+    void getLottoPublishCount() {
+        int lottoPublishCount = purchase.getLottoPublishCount();
+
+        Assertions.assertThat(lottoPublishCount).isEqualTo(2);
+    }
+
+    @DisplayName("구입 금액 숫자 예외 처리 테스트")
+    @Nested
+    class NumberException {
+
+        @DisplayName("구입 금액이 1000원 단위가 아닌 경우 예외를 터트린다.")
+        @Test
+        void isDividedByLottoAmount() {
+            int purchase = 11003;
+
+            Assertions.assertThatThrownBy( () -> Purchase.create(purchase))
+                    .isInstanceOf(IllegalArgumentException.class)
+                    .hasMessageContaining(InputException.MONEY_NOT_DIVIDE_BY_LOTTO_UNIT.message());
+        }
+    }
+
+    @DisplayName("구입 금액 문자열 예외 처리 테스트")
+    @Nested
+    class StringException {
+
+        @DisplayName("구입 금액이 숫자가 아닌 경우 예외를 터트린다.")
+        @Test
+        void hasDigitsOnly() {
+            String purchase = "3!@# A";
+            System.setIn(generateUserInput(purchase));
+
+            Assertions.assertThatThrownBy(InputView::inputLottoPurchaseAmount)
+                    .isInstanceOf(IllegalArgumentException.class)
+                    .hasMessageContaining(InputException.MONEY_NOT_DIGIT.message());
+        }
+    }
+
+    public static InputStream generateUserInput(String input) {
+        return new ByteArrayInputStream(input.getBytes());
+    }
+}
